Adding localization-expansion (not pseudo)
Designers' feedback: even if pseudo-language is “readable” and helps
with clippings for the ascenders\descenders, the most important issue to
address is about truncations happening after translations happen, just
because of lack of space in the UI redlines. To address that, what
Designers lack is “just” the knowledge of the potential expansion-rates
applied.